+++ solutions.py
@@ -169,65 +169,6 @@ class LongestCommonPrefix:
         return shortest_word
 
 
-class ValidParentheses:
-    @staticmethod
-    def my_solution(s: str) -> bool:
-        """
-        LeetCode Question Nr.20
-        Given a string s containing just the characters '(', ')', '{', '}', '[' and ']', determine if the input string is valid.
-        An input string is valid if:
-            Open brackets must be closed by the same type of brackets.
-            Open brackets must be closed in the correct order.
-
-        Input: s = "{[]}"  Output: true
-        Input: s = "([)]"  Output: false
-        Input: s = "()[]{}"  Output: true
-        """
-
-        ## technic used: stack
-
-        if len(s) <= 1:
-            return False
-
-        stack = []
-        end_bracket_dict = {"]":"[", "}":"{", ")":"("}
-
-        for char in s:
-            if char in end_bracket_dict.values():
-                stack.append(char)
-            elif char in end_bracket_dict.keys():
-                if len(stack) == 0 or end_bracket_dict[char] != stack.pop():
-                    return False
-            else:
-                return False
-        return stack == []   # to avoid something like "(("
-    
-    @staticmethod
-    def great_ans1(s: str) -> bool:
-        """
-        https://leetcode.com/problems/valid-parentheses/discuss/316753/Python-4ms-Faster-then-100-with-explanation
-        """
-
-        ## technic used: stack
-
-        if len(s) <= 1:
-            return False
-
-        stack = []
-        start_bracket_dict = {"[":"]", "{":"}", "(":")"}
-
-        for char in s:
-            if char in start_bracket_dict.keys():
-                stack.append(char)
-            elif char in start_bracket_dict.values():
-                if len(stack) == 0 or char != start_bracket_dict[stack.pop()]:
-                    return False
-            else:
-                return False
-        
-        return stack == []
-
-
 class MergeTwoSortedLists:
     @staticmethod
     def my_solution(l1: ListNode, l2: ListNode) -> ListNode:
@@ -261,97 +202,7 @@ class MergeTwoSortedLists:
 
         
 
-class BinarySearch:
-    """
-    Given an array of integers nums which is sorted in ascending order, 
-    and an integer target, write a function to search target in nums. 
-    If target exists, then return its index. Otherwise, return -1.
-    You must write an algorithm with O(log n) runtime complexity.
 
-    Input: nums = [-1,0,3,5,9,12], target = 9
-    Output: 4
-    Explanation: 9 exists in nums and its index is 4
-    """
-
-    @staticmethod
-    def my_solution(nums: List[int], target: int) -> int:
-        """
-        tech use : binary search
-        """
-        target_index = 0
-        
-        while len(nums) > 1:
-            split_index = int(len(nums)/2)
-            
-            if target < nums[split_index]:
-                nums = nums[:split_index]
-                target_index += 0 
-            elif target > nums[split_index]:
-                nums = nums[split_index:]
-                target_index += split_index
-            else:
-                target_index += split_index
-                return target_index
-        
-        return target_index if nums[0] == target else -1
-            
-
-    @staticmethod
-    def great_ans1(nums: List[int], target: int) -> int:
-        """
-        tech use : binary search
-        https://www.youtube.com/watch?v=s4DPM8ct1pI
-        """
-        l, r = 0, len(nums) - 1
-
-        while l <= r:
-            mid = (l + r) // 2  # to prevent overfloat: mid = l + (r-l)//2
-            if target < nums[mid]:
-                r = mid - 1
-            elif target > nums[mid]:
-                l = mid + 1
-            else:
-                return mid
-        return -1    
-
-
-class SingleNumber:
-    """
-    Given a non-empty array of integers nums, every element appears twice except for one. Find that single one.
-    You must implement a solution with a linear runtime complexity and use only constant extra space.
-
-    Input: nums = [4,1,2,1,2]
-    Output: 4
-    """
-
-    @staticmethod
-    def my_solution(nums: List[int]) -> int:
-        i = 0
-        while i < len(nums)-1:
-            if nums[i] not in nums[:i] and nums[i] not in nums[i+1:]:
-                return nums[i]
-            i += 1
-        return nums[i]
-            
-
-    @staticmethod
-    def great_ans1(nums: List[int]) -> int:
-        """
-        tech use : bit manipulation / XOR
-        https://leetcode.com/problems/single-number/discuss/1771771/Think-it-through-oror-Time%3A-O(n)-Space%3A-O(1)-oror-Python-Explained
-
-        XOR:
-        0^0 = 0 ; 1^1 = 0 ; 1^0 = 1
-        2^2 = 0 ; 2^3 = 5 
-        A^B^A = A^A^B : 1^2^3^2^3 = 1 ^ (2^2) ^ (3^3) = 1^0^0 = 1
-        https://accu.org/journals/overload/20/109/lewin_1915/
-        """
-        
-        res = 0
-        for num in nums:
-            res = res ^ num
-        
-        return res
 
 
 class SameTree:
